Recipe_maker/requirements.txt
google-generativeai
pandas
numpy
chromadb
sentence-transformers
streamlit 
torch 
transformers 
diffusers
accelerate 
pillow 
requests